refactor: centralize request IP resolver in core (#10216)
