+++ packages/api-key/src/utils.ts
@@ -12,47 +12,3 @@ export function isAPIError(error: unknown): error is APIError {
 		(error as any)?.name === "APIError"
 	);
 }
-
-import type { BetterAuthOptions } from "@better-auth/core";
-import { isDevelopment, isTest } from "@better-auth/core/env";
-import { getIPFromHeader } from "@better-auth/core/utils/ip";
-
-const LOCALHOST_IP = "127.0.0.1";
-
-// TODO: this duplicates `better-auth` request-IP resolver. Centralize it
-// behind a shared export so the security-sensitive logic lives in one place.
-export function getIp(
-	req: Request | Headers,
-	options: BetterAuthOptions,
-): string | null {
-	if (options.advanced?.ipAddress?.disableIpTracking) {
-		return null;
-	}
-
-	const headers = "headers" in req ? req.headers : req;
-
-	const defaultHeaders = ["x-forwarded-for"];
-
-	const ipHeaders =
-		options.advanced?.ipAddress?.ipAddressHeaders || defaultHeaders;
-
-	for (const key of ipHeaders) {
-		const value = "get" in headers ? headers.get(key) : headers[key];
-		if (typeof value === "string") {
-			const ip = getIPFromHeader(value, {
-				ipv6Subnet: options.advanced?.ipAddress?.ipv6Subnet,
-				trustedProxies: options.advanced?.ipAddress?.trustedProxies,
-			});
-			if (ip) {
-				return ip;
-			}
-		}
-	}
-
-	// Fallback to localhost IP in development/test environments when no IP found in headers
-	if (isTest() || isDevelopment()) {
-		return LOCALHOST_IP;
-	}
-
-	return null;
-}
